requirements.txt
Django==4.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2021.1
sqlparse==0.4.1